Fixes for other testcases in testNG
