+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch04/cookies/DeleteCookiesNGTest.java
@@ -18,16 +18,13 @@ package io.github.bonigarcia.webdriver.testng.ch04.cookies;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.time.Duration;
 import java.util.Set;
-
+import org.openqa.selenium.WebDriver.Options;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Cookie;
-import org.testng.annotations.*;
 import io.github.bonigarcia.webdriver.HelperClass.TestSetup;
 
 public class DeleteCookiesNGTest extends TestSetup {
-
     public void testDeleteCookieNamedUsername() {
         driver.get(
                 "https://bonigarcia.dev/selenium-webdriver-java/cookies.html");
