+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch04/cookies/EditCookiesNGTest.java
@@ -19,7 +19,7 @@ package io.github.bonigarcia.webdriver.testng.ch04.cookies;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.Duration;
-
+import org.openqa.selenium.WebDriver.Options;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Cookie;
 import org.testng.annotations.*;
