+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch04/dialogs/ConfirmNGTest.java
@@ -20,13 +20,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.Duration;
 import org.testng.annotations.*;
+
+import io.github.bonigarcia.webdriver.HelperClass.TestSetup;
+
 import org.openqa.selenium.Alert;
 import org.openqa.selenium.By;
 
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class ConfirmNGTest {
+public class ConfirmNGTest extends TestSetup {
 
     @Test
     public void testConfirmDialogBox() {
