+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch04/downdown/DataListNGTest.java
@@ -23,6 +23,7 @@ import java.time.Duration;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
 import org.testng.annotations.*;
 import io.github.bonigarcia.webdriver.HelperClass.TestSetup;
