+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch05/cdp/DeviceEmulationNGTest.java
@@ -28,7 +28,7 @@ import java.util.Optional;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.devtools.DevTools;
-import org.openqa.selenium.devtools.v106.network.Network;
+import org.openqa.selenium.devtools.v111.network.Network;
 import org.slf4j.Logger;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
