+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch05/cdp/ManageCookiesNGTest.java
@@ -30,8 +30,8 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.devtools.DevTools;
-import org.openqa.selenium.devtools.v106.network.Network;
-import org.openqa.selenium.devtools.v106.network.model.Cookie;
+import org.openqa.selenium.devtools.v111.network.Network;
+import org.openqa.selenium.devtools.v111.network.model.Cookie;
 import org.slf4j.Logger;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
@@ -41,55 +41,55 @@ import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class ManageCookiesNGTest {
 
-    static final Logger log = getLogger(lookup().lookupClass());
+        static final Logger log = getLogger(lookup().lookupClass());
 
-    WebDriver driver;
+        WebDriver driver;
 
-    DevTools devTools;
+        DevTools devTools;
 
-    @BeforeMethod
-    public void setup() {
-        driver = WebDriverManager.chromedriver().create();
-        devTools = ((ChromeDriver) driver).getDevTools();
-        devTools.createSession();
-    }
+        @BeforeMethod
+        public void setup() {
+                driver = WebDriverManager.chromedriver().create();
+                devTools = ((ChromeDriver) driver).getDevTools();
+                devTools.createSession();
+        }
 
-    @AfterMethod
-    public void teardown() throws InterruptedException {
-        // FIXME: pause for manual browser inspection
-        Thread.sleep(Duration.ofSeconds(3).toMillis());
+        @AfterMethod
+        public void teardown() throws InterruptedException {
+                // FIXME: pause for manual browser inspection
+                Thread.sleep(Duration.ofSeconds(3).toMillis());
 
-        devTools.close();
-        driver.quit();
-    }
+                devTools.close();
+                driver.quit();
+        }
 
-    @Test
-    public void testManageCookies() {
-        devTools.send(Network.enable(Optional.empty(), Optional.empty(),
-                Optional.empty()));
-        driver.get(
-                "https://bonigarcia.dev/selenium-webdriver-java/cookies.html");
+        @Test
+        public void testManageCookies() {
+                devTools.send(Network.enable(Optional.empty(), Optional.empty(),
+                                Optional.empty()));
+                driver.get(
+                                "https://bonigarcia.dev/selenium-webdriver-java/cookies.html");
 
-        // Read cookies
-        List<Cookie> cookies = devTools.send(Network.getAllCookies());
-        cookies.forEach(cookie -> log.debug("{}={}", cookie.getName(),
-                cookie.getValue()));
-        List<String> cookieName = cookies.stream()
-                .map(cookie -> cookie.getName()).sorted()
-                .collect(Collectors.toList());
-        Set<org.openqa.selenium.Cookie> seleniumCookie = driver.manage()
-                .getCookies();
-        List<String> selCookieName = seleniumCookie.stream()
-                .map(selCookie -> selCookie.getName()).sorted()
-                .collect(Collectors.toList());
-        assertThat(cookieName).isEqualTo(selCookieName);
+                // Read cookies
+                List<Cookie> cookies = devTools.send(Network.getAllCookies());
+                cookies.forEach(cookie -> log.debug("{}={}", cookie.getName(),
+                                cookie.getValue()));
+                List<String> cookieName = cookies.stream()
+                                .map(cookie -> cookie.getName()).sorted()
+                                .collect(Collectors.toList());
+                Set<org.openqa.selenium.Cookie> seleniumCookie = driver.manage()
+                                .getCookies();
+                List<String> selCookieName = seleniumCookie.stream()
+                                .map(selCookie -> selCookie.getName()).sorted()
+                                .collect(Collectors.toList());
+                assertThat(cookieName).isEqualTo(selCookieName);
 
-        // Clear cookies
-        devTools.send(Network.clearBrowserCookies());
-        List<Cookie> cookiesAfterClearing = devTools
-                .send(Network.getAllCookies());
-        assertThat(cookiesAfterClearing).isEmpty();
+                // Clear cookies
+                devTools.send(Network.clearBrowserCookies());
+                List<Cookie> cookiesAfterClearing = devTools
+                                .send(Network.getAllCookies());
+                assertThat(cookiesAfterClearing).isEmpty();
 
-        driver.findElement(By.id("refresh-cookies")).click();
-    }
+                driver.findElement(By.id("refresh-cookies")).click();
+        }
 }
