+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch05/cdp/PerformanceMetricsNGTest.java
@@ -27,8 +27,8 @@ import java.util.Optional;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.devtools.DevTools;
-import org.openqa.selenium.devtools.v106.performance.Performance;
-import org.openqa.selenium.devtools.v106.performance.model.Metric;
+import org.openqa.selenium.devtools.v111.performance.Performance;
+import org.openqa.selenium.devtools.v111.performance.model.Metric;
 import org.slf4j.Logger;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
