+++ selenium-webdriver-testng/src/test/java/io/github/bonigarcia/webdriver/testng/ch07/tests/LoginNGTest.java
@@ -19,14 +19,13 @@ package io.github.bonigarcia.webdriver.testng.ch07.tests;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
+import io.github.bonigarcia.webdriver.HelperClass.TestSetup;
 import io.github.bonigarcia.webdriver.testng.ch07.page_objects.LoginPage;
 
-public class LoginNGTest {
+public class LoginNGTest extends TestSetup {
+
     @Test
     public void testLoginSuccess() {
         LoginPage login = new LoginPage(driver);
